feat: define standard schema adapter in form-core
