+++ packages/form-core/src/index.ts
@@ -5,3 +5,4 @@ export * from './util-types'
 export * from './types'
 export * from './mergeForm'
 export * from './formOptions'
+export * from './standardSchemaValidator'
