+++ packages/form-core/tests/FormApi.spec.ts
@@ -1,5 +1,4 @@
 import { describe, expect, it, vi } from 'vitest'
-import * as v from 'valibot'
 import { FieldApi, FormApi } from '../src/index'
 import { sleep } from './utils'
 
@@ -1614,68 +1613,6 @@ describe('form api', () => {
     expect(field.getMeta().errorMap.onChange).toEqual('first name is required')
     await form.validateField('firstName', 'change')
     expect(field.getMeta().errorMap.onChange).toEqual('first name is required')
-  })
-
-  it('should support StandardSchema sync validation with valibot', async () => {
-    const form = new FormApi({
-      defaultValues: {
-        firstName: '',
-        lastName: '',
-      },
-      validators: {
-        onChange: v.object({
-          firstName: v.pipe(
-            v.string(),
-            v.minLength(3, 'First name is too short'),
-          ),
-          lastName: v.string(),
-        }),
-      },
-    })
-
-    const field = new FieldApi({
-      form,
-      name: 'firstName',
-    })
-
-    field.mount()
-
-    field.setValue('')
-    expect(form.state.errors).toStrictEqual(['First name is too short'])
-  })
-
-  it('should support StandardSchema async validation with valibot', async () => {
-    vi.useFakeTimers()
-
-    const form = new FormApi({
-      defaultValues: {
-        firstName: '',
-        lastName: '',
-      },
-      validators: {
-        onChangeAsync: v.objectAsync({
-          firstName: v.pipeAsync(
-            v.string(),
-            v.checkAsync(async (val) => {
-              await sleep(1)
-              return val.length > 3
-            }, 'First name is too short'),
-          ),
-          lastName: v.string(),
-        }),
-      },
-    })
-
-    const field = new FieldApi({
-      form,
-      name: 'firstName',
-    })
-
-    field.mount()
-
-    field.setValue('')
-    await vi.runAllTimersAsync()
-    expect(form.state.errors).toStrictEqual(['First name is too short'])
   })
 
   it('should show onSubmit errors', async () => {
